refactor: consolidate CapitalisationPolicy enum and capitalise() helper
- Extract shared CapitalisationPolicy enum to capitalisation/mod.rs
- Add from_config() method to deduplicate configure() parsing in CP01/CP03
- Move capitalise() to utils.rs (was duplicated in CP01 and CP03)
